pyproject.toml
[project]
name = "chinook"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.8",
    "drf-spectacular>=0.28.0",
    "psycopg[binary]>=3.2.10",
]
